requirements.txt
requests
python-dotenv
urllib3
python-dateutil
colorama==0.4.6

# Pandas + wsparcie dla write_pandas
pandas
pyarrow
snowflake-connector-python[pandas]
